Revert "refactor: dont use any stdlib headers"
This reverts commit cbb0c3aadd301e9a06957a661bceb7f5361ef5ad.

+++ kernel/interrupts.c
@@ -4,6 +4,7 @@
 
 #include <my_stdio.h>
 #include <my_stdlib.h>
+#include <stddef.h>
 #include <stdint.h>
 #include <vga_color.h>
 
@@ -67,7 +68,7 @@ void exc_handler(int vector, struct pushad_frame *registers, struct interrupt_fr
 
     uint64_t *rsp        = (void *)interrupts->rsp;
     uint64_t  stack_top  = (uint64_t)get_stack_top();
-    uint32_t  stack_size = (stack_top - (uint64_t)rsp) / sizeof(long);
+    uint32_t  stack_size = (stack_top - (uint64_t)rsp) / sizeof(size_t);
 
     if (stack_size > STACK_DUMP_ROWS * STACK_DUMP_COLUMNS) {
         stack_size = STACK_DUMP_ROWS * STACK_DUMP_COLUMNS;
